refactor: rename utils/linters utils/cli-tools
since it's now used for babel, it doesn't make sense to name it "linters"
